Switching to provider + singleton approach
